backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
cohere==4.3.1
qdrant-client>=1.7.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pytest==7.4.3
slowapi==0.1.9
limits==5.6.0
PyPDF2==3.0.1
ebooklib==0.18
beautifulsoup4==4.12.2
python-multipart==0.0.6